set errors in account analysis on facility changes
