+++ src/app/account/account-home/account-home-summary/account-home-summary.component.html
@@ -133,6 +133,10 @@
                     <app-calculating-spinner *ngIf="calculating != false" [message]="'Calculating Monthly Results...'"
                         [error]="calculating == 'error'">
                     </app-calculating-spinner>
+                    <div class="alert alert-danger"
+                        *ngIf="latestAnalysisItem.setupErrors.hasError || latestAnalysisItem.setupErrors.facilitiesSelectionsInvalid">
+                        Errors found in selected analysis.
+                    </div>
                 </ng-container>
             </div>
         </div>
