+++ src/app/facility/analysis/analysis.component.ts
@@ -1,14 +1,13 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
-import { Subscription, firstValueFrom } from 'rxjs';
+import { Subscription } from 'rxjs';
 import { FacilitydbService } from 'src/app/indexedDB/facility-db.service';
 import { UtilityMeterDatadbService } from 'src/app/indexedDB/utilityMeterData-db.service';
 import { UtilityMeterGroupdbService } from 'src/app/indexedDB/utilityMeterGroup-db.service';
-import { IdbAccount, IdbAccountAnalysisItem, IdbAnalysisItem, IdbFacility, IdbUtilityMeterData, IdbUtilityMeterGroup } from 'src/app/models/idb';
+import { IdbAnalysisItem, IdbFacility, IdbUtilityMeterData, IdbUtilityMeterGroup } from 'src/app/models/idb';
 import { AnalysisService } from './analysis.service';
 import { AnalysisDbService } from 'src/app/indexedDB/analysis-db.service';
 import { AccountAnalysisDbService } from 'src/app/indexedDB/account-analysis-db.service';
-import { AnalysisValidationService } from 'src/app/shared/helper-services/analysis-validation.service';
 
 @Component({
   selector: 'app-analysis',
@@ -28,8 +27,7 @@ export class AnalysisComponent implements OnInit {
     private facilityDbService: FacilitydbService,
     private analysisService: AnalysisService,
     private analysisDbService: AnalysisDbService,
-    private accountAnalysisDbService: AccountAnalysisDbService,
-    private analysisValidationService: AnalysisValidationService) { }
+    private accountAnalysisDbService: AccountAnalysisDbService) { }
 
   ngOnInit(): void {
     this.utilityMeterDataSub = this.utilityMeterDataDbService.facilityMeterData.subscribe(val => {
@@ -63,19 +61,6 @@ export class AnalysisComponent implements OnInit {
   }
 
   async updateAccountValidation(allAnalysisItems: Array<IdbAnalysisItem>) {
-    let accountAnalysisItems: Array<IdbAccountAnalysisItem> = this.accountAnalysisDbService.accountAnalysisItems.getValue();
-    let hasChanges: boolean = false;
-    for (let i = 0; i < accountAnalysisItems.length; i++) {
-      let item: IdbAccountAnalysisItem = accountAnalysisItems[i];
-      let results: { analysisItem: IdbAccountAnalysisItem, isChanged: boolean } = this.analysisValidationService.updateFacilitySelectionErrors(item, allAnalysisItems);
-      if (results.isChanged) {
-        accountAnalysisItems[i] = results.analysisItem;
-        await firstValueFrom(this.accountAnalysisDbService.updateWithObservable(accountAnalysisItems[i]));
-        hasChanges = true;
-      }
-    }
-    if (hasChanges) {
-      this.accountAnalysisDbService.accountAnalysisItems.next(accountAnalysisItems);
-    }
+    await this.accountAnalysisDbService.updateAccountValidation(allAnalysisItems);
   }
 }
