+++ src/app/facility/utility-data/predictor-data/predictor-data.component.html
@@ -3,151 +3,4 @@
     <div class="content-padding">
         <router-outlet></router-outlet>
     </div>
-
-
-
-    <!-- <h4>Predictors</h4>
-
-    <div *ngIf="!hasData" class="text-center mt-5">
-        <h5>No predictors found for this facility.</h5>
-        <p>Please <i class="fa fa-plus"></i> Add New Predictor or <i class="fa fa-upload"></i> Import multiple
-            predictors from
-            excel.</p>
-        <div>
-            <button class="btn btn-primary me-2" (click)="editPredictors()"><i class="fa fa-plus"></i> Add New
-                Predictor</button>
-            <button class="btn btn-outline" (click)="uploadData()"><i class="fa fa-upload"></i> Import Data</button>
-        </div>
-    </div>
-
-    <ng-container *ngIf="hasData">
-
-        <div class="d-flex w-100 justify-content-between mb-2">
-            <div>
-                <button class="btn action-btn" (click)="editPredictors()">Add/Edit Predictors</button>
-            </div>
-
-            <div class="d-flex">
-                <!--Bulk Delete--
-                <div class="pe-2 bulk-delete" [class.show]="hasCheckedItems">
-                    <button class="btn btn-danger" (click)='openBulkDelete()' [disabled]="!hasCheckedItems">Bulk
-                        Delete</button>
-                </div>
-                <!--Items Per Page--
-                <app-table-items-dropdown></app-table-items-dropdown>
-                <div class="ps-2">
-                    <button class="btn nav-btn" (click)="uploadData()"><i class="fa fa-upload"></i> Import
-                        Data</button>
-                </div>
-
-
-                <div class="ps-2">
-                    <button class="btn action-btn" (click)="addPredictorEntry()"><i class="fa fa-plus"></i> Add
-                        Predictor
-                        Entry</button>
-                </div>
-            </div>
-        </div>
-
-        <table class="table utility-data table-sm table-bordered table-hover" #predictorTable [ngClass]="{'copying-table': copyingTable}">
-            <thead class="sortable">
-                <tr>
-                    <th class="input" *ngIf="!copyingTable">
-                        <input #masterCheckbox type="checkbox" (change)="checkAll()" [(ngModel)]="allChecked">
-                    </th>
-                    <th (click)="setOrderDataField('date')" [ngClass]="{'active': orderDataField == 'date'}">Date
-                    </th>
-                    <th class="header" *ngFor="let predictor of facilityPredictors; let predictorIndex = index;"
-                        (click)="setOrderDataField(predictor.name)"
-                        [ngClass]="{'active': orderDataField == predictor.name}">
-                        <span *ngIf="predictor.production" class="fa fa-check-circle-o"></span>
-                        {{predictor.name}}
-                        <br *ngIf="predictor.unit">
-                        <span class="small block" *ngIf="predictor.unit">({{predictor.unit}})</span>
-                    </th>
-                    <td class="actions" *ngIf="!copyingTable">
-
-                    </td>
-                </tr>
-            </thead>
-            <tbody class="table-group-divider">
-                <tr
-                    *ngFor="let predictorEntry of facilityPredictorEntries | predictorsOrderBy: orderDataField: orderByDirection | slice: ((currentPageNumber-1) * itemsPerPage): currentPageNumber * itemsPerPage">
-                    <td class="input" *ngIf="!copyingTable">
-                        <input type="checkbox" name="predictorEntry" (change)="setHasChecked()"
-                            [(ngModel)]="predictorEntry.checked">
-                    </td>
-                    <td>{{predictorEntry.date | date:'MMM, yyyy'}}</td>
-                    <td *ngFor="let predictor of predictorEntry.predictors">{{predictor.amount | customNumber}}</td>
-                    <td class="actions" *ngIf="!copyingTable">
-                        <div class="btn-group">
-                            <button class="btn btn-sm btn-outline"  (click)="setEditPredictorEntry(predictorEntry)"title="Edit">
-                                <i class="fa fa-pencil"></i>
-                            </button>
-                            <button class="btn btn-sm btn-outline"  (click)="setDeletePredictorEntry(predictorEntry)" title="Delete">
-                                <i class="fa fa-trash"></i>
-                            </button>
-                        </div>
-                    </td>
-                </tr>
-            </tbody>
-        </table>
-        <div class="d-flex w-100 justify-content-between">
-            <div>
-                <button class="btn action-btn btn-sm" (click)="copyTable()">
-                    <span class="fa fa-copy"></span> Copy Table
-                </button>
-            </div>
-            <ngb-pagination [collectionSize]="facilityPredictorEntries.length" [(page)]="currentPageNumber"
-                [pageSize]="itemsPerPage" [maxSize]="5" [rotate]="true" aria-label="Pagination">
-            </ngb-pagination>
-        </div>
-    </ng-container> -->
 </div>
-<!-- 
-<div class="window" [class.open]="predictorEntryToEdit">
-    <div class="windowContent">
-        <app-edit-predictor-entry-row *ngIf="predictorEntryToEdit" (emitClose)="cancelEditPredictorEntry()"
-            [predictorEntry]="predictorEntryToEdit" [addOrEdit]="addOrEdit">
-        </app-edit-predictor-entry-row>
-    </div>
-</div>
-
-
-<div class="window" [class.open]="showEditPredictors">
-    <div class="windowContent windowContentLg">
-        <app-edit-predictors *ngIf="showEditPredictors" (emitClose)="closeEditPredictors()"></app-edit-predictors>
-    </div>
-</div>
-
-
-<div [ngClass]="{'windowOverlay': predictorEntryToDelete || showBulkDelete || showEditPredictors || predictorEntryToEdit}"></div>
-<div class="popup" [class.open]="predictorEntryToDelete">
-    <div class="popup-header">Delete Predictor
-        <button class="item-right" (click)="cancelDeletePredictorEntry()">x</button>
-    </div>
-    <div class="popup-body">
-        <p>WARNING: By deleting this predictor, you are also deleting all of the data associated with it. Are you sure
-            you want to preform this action?</p>
-    </div>
-    <div class="saveCancel popup-footer text-end">
-        <button class="btn btn-secondary" (click)="cancelDeletePredictorEntry()">Cancel</button>
-        <button class="btn btn-danger" (click)="confirmDeletePredictorEntry()">Delete</button>
-    </div>
-</div>
-
-
-<!--bulk delete--
-<div class="popup" [class.open]="showBulkDelete">
-    <div class="popup-header" *ngIf="showBulkDelete">Delete Predictor Entries
-        <button class="item-right" (click)="cancelBulkDelete()">x</button>
-    </div>
-    <div class="popup-body">
-        <p>WARNING: Deleting predictor entries cannot be undone.
-            Are you sure you want to preform this action?</p>
-    </div>
-    <div class="saveCancel popup-footer text-end">
-        <button class="btn btn-secondary" (click)="cancelBulkDelete()">Cancel</button>
-        <button class="btn btn-danger" (click)="bulkDelete()">Delete</button>
-    </div>
-</div> -->
