+++ src/app/indexedDB/db-changes.service.ts
@@ -95,6 +95,15 @@ export class DbChangesService {
 
   async setAccountAnalysisItems(account: IdbAccount) {
     let accountAnalysisItems: Array<IdbAccountAnalysisItem> = await this.accountAnalysisDbService.getAllAccountAnalysisItems(account.guid);
+    let facilityAnalysisItems: Array<IdbAnalysisItem> = this.analysisDbService.accountAnalysisItems.getValue();
+    for (let i = 0; i < accountAnalysisItems.length; i++) {
+      let updateAnalysis: { analysisItem: IdbAccountAnalysisItem, isChanged: boolean } = this.updateDbEntryService.updateAccountAnalysis(accountAnalysisItems[i], account, facilityAnalysisItems);
+      if (updateAnalysis.isChanged) {
+        accountAnalysisItems[i] = updateAnalysis.analysisItem;
+        await firstValueFrom(this.accountAnalysisDbService.updateWithObservable(accountAnalysisItems[i]));
+      };
+    }
+   
     this.accountAnalysisDbService.accountAnalysisItems.next(accountAnalysisItems);
   }
 
@@ -107,7 +116,6 @@ export class DbChangesService {
         await firstValueFrom(this.analysisDbService.updateWithObservable(analysisItems[i]));
       };
     }
-    console.log('SETT')
     this.analysisDbService.accountAnalysisItems.next(analysisItems);
     if (facility) {
       this.setFacilityAnalysisItems(facility);
